--- GyroData/Model/AnalysisType.swift
@@ -1,30 +1,0 @@
-//
-//  AnalysisType.swift
-//  GyroData
-//
-//  Created by unchain, Ellen J, yeton on 2022/12/26.
-
-import Foundation
-
-enum AnalysisType: String, Codable {
-    case accelerate = "accelerate"
-    case gyroscope = "gyroscope"
-}
-
-struct CellModel: Codable {
-    let id: UUID
-    let analysisType: String
-    let savedAt: Date
-    let measurementTime: Double
-}
-
-class EnvironmentGraphModel: ObservableObject {
-    @Published var graphModels: [GraphModel] = []
-}
-
-struct GraphModel: Codable {
-    let x: Double
-    let y: Double
-    let z: Double
-    let measurementTime: Double
-}
